:wrench: work without jquery
rah FSA ratings are back

down to 974 bytes now, not as small as it could be
is duplicating gU.get() but adds in headers, which
I don't want to fatten gU.get up with yet. This script
is post loaded so it's ok that it's a bit fatter.

i could shave a load of this with a better minifier...