:mouse: lose the getElementsByTagName('b') - 1555 bytes
Changed Gigantasaur to make wiki links <a>like this</a>
instead of <b>like this</b> which means I only cycle through one
kind of tag and there's no inserting of new elements now

Is it an SEO risk? http://www.clarkeology.com/m/24229
Not really because I have no seo to speak of anyway so better
to be honest, and get the page size down and speed up rather
than have a load of artificial bold tags. Or was the issue that
I had too many internal links? Will see if it makes a difference..